Разнесение сервема по дирректориям
