+++ homework-3/css/style.css
@@ -56,4 +56,3 @@ h1 {
     box-shadow: 5px 10px 5px grey;
 
 }
-
