+++ homework-3/www/templates/index.html
@@ -2,7 +2,6 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <!-- <link rel="stylesheet" type="text/css" href="style.css"> -->
     <title>{{.Name}}</title>
     {{template "head"}}
 </head>
